S2 P1 time_control fix8
Sprint 2 Part 1 time_control fix8

+++ sprint2/problems/time_control/solution/src/model.h
@@ -305,9 +305,9 @@ public:
 private:
     void SetStartPosition();
 
-    void MakeHorizontalMovement(const int64_t& tick);
+    void MakeHorizontalMovement(const double& tick);
 
-    void MakeVerticalMovement(const int64_t& tick);
+    void MakeVerticalMovement(const double& tick);
 
     std::string name_;
     DogCoords position_;
